Add files via upload
Backdoor attack is a new security threat for
the large language models. In this project, we implemented
a word based backdoor attack on large language model,
and a perplexity defense system to defend the attack. During
poisoning we compromised 1%, 3%, 5%, 10%, 15%,
and 20% of the training datasets and varied the trigger
length from 1 to 3. From the experimental results, it can be
seen that the attack is successful for all the poisoning and
trigger length conditions. The defense system is effective
when the trigger length is 1 but it starts to deteriorate
when we increase the length.